+++ clang-tools-extra/test/clang-tidy/checkers/bugprone/avoid-dangling-calls-custom.cpp
@@ -1,0 +1,2 @@
+// RUN: %check_clang_tidy %s bugprone-avoid-dangling-calls %t
+